requirements.txt
Django>=4.2.0
django-cors-headers>=4.0.0
joblib>=1.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyserial>=3.5